Split setup functions into 2 files
